Fix header and create new search component
