+++ audit/ask-cal/source/components/header/Header.module.css
@@ -1,8 +1,15 @@
 .header {
-  grid-area: header;
   display: flex;
   justify-content: space-between;
   align-items: center;
-  padding: 16px;
-  border-bottom: var(--palette-gray-dark-2) 1px solid;
+  margin-top: 20%;
+}
+
+.header_results {
+  display: flex;
+  justify-content: space-between;
+  align-items: center;
+  padding: 2em 3em;
+  width: 100%;
+  border-bottom: var(--palette-gray-light-2) solid 1px;
 }
