+++ audit/ask-cal/source/components/logoblock/LogoBlock.module.css
@@ -4,5 +4,5 @@
 }
 
 .logo_block svg {
-  padding-left: 1em;
+  padding-right: 1em;
 }
