+++ audit/ask-cal/source/pages/home/HomePage.module.css
@@ -4,21 +4,4 @@
   align-items: center;
   justify-content: center;
   width: 100vw;
-  height: 80vh;
-}
-
-.facet_container {
-  display: flex;
-  justify-content: space-between;
-  margin-top: 8em;
-  width: 50vw;
-}
-
-.facet_container > div {
-  width: 30%;
-}
-
-.search_input {
-  width: 50vw;
-  margin-top: 3em;
 }
