+++ audit/ask-cal/source/pages/results/home/ResultsPage.module.css
@@ -6,22 +6,6 @@
   height: 100vh;
 }
 
-header {
-  display: flex;
-  gap: 2em;
-  align-items: center;
-  margin: 4em 0 3em 0;
-  padding: 0.5em 0;
-}
-
-header > form {
-  flex: 2;
-}
-
-header > div {
-  flex: 3;
-}
-
 .horizontal_container {
   display: flex;
   width: 100%;
@@ -43,17 +27,6 @@ header > div {
   flex-direction: column;
   margin: 2em 1em;
   max-width: 55%;
-}
-
-.facet_container {
-  display: flex;
-  flex-wrap: wrap;
-  justify-content: space-between;
-  width: 100%;
-}
-
-.facet_container > div {
-  width: 30%;
 }
 
 .results {
