+++ audit/ask-cal/source/pages/results/home/ResultsPage.tsx
@@ -1,163 +1,24 @@
 import styles from "./ResultsPage.module.css";
 import { useState } from "react";
 
-import { SearchInput } from "@leafygreen-ui/search-input";
-import { Combobox, ComboboxOption } from "@leafygreen-ui/combobox";
 import Card from "@leafygreen-ui/card";
 import Code from "@leafygreen-ui/code";
 import IconButton from "@leafygreen-ui/icon-button";
 import Button from "@leafygreen-ui/button";
 import Icon from "@leafygreen-ui/icon";
-import { FacetGroup, Facet, CodeExample } from "../../../constants/types";
+import { Body, H2, H3 } from "@leafygreen-ui/typography";
+import { CodeExample } from "../../../constants/types";
 
 import { useAcala } from "../../../providers/UseAcala";
 
-import {
-  CodeExampleDisplayValues,
-  CodeExampleCategory,
-} from "../../../constants/categories";
-import { DocsSetDisplayValues, DocsSet } from "../../../constants/docsSets";
-import {
-  ProgrammingLanguageDisplayValues,
-  ProgrammingLanguage,
-} from "../../../constants/programmingLanguages";
-import { Body, H2, H3 } from "@leafygreen-ui/typography";
-
 function Resultspage() {
-  const [searchQuery, setSearchQuery] = useState<string>("");
-  const [facets, setFacets] = useState<FacetGroup>({
-    programmingLanguage: "",
-    category: "",
-    docsSet: "",
-  });
   const [selectedCodeExample, setSelectedCodeExample] =
     useState<CodeExample | null>(null);
 
-  const { search, results } = useAcala();
-
-  // TODO: this also exists on the homepage. Move to a common place.
-  const handleSearch = async (event: React.FormEvent<HTMLFormElement>) => {
-    setSearchQuery("");
-    // Get the value from the input element. Look for the role "search".
-    const inputElement = event.currentTarget.querySelector(
-      "input"
-    ) as HTMLInputElement;
-    const value = inputElement.value;
-
-    if (!value) {
-      console.error("Search input is empty");
-      return;
-    }
-
-    setSearchQuery(value);
-
-    // Mock search, as getting CORS errors
-    // TODO: make this work for real
-    await search({
-      bodyContent: {
-        queryString: searchQuery,
-        LanguageFacet: facets.programmingLanguage,
-        CategoryFacet: facets.category,
-        docsSet: facets.docsSet,
-      },
-      mock: true,
-    });
-  };
-
-  const handleFacetChange = ({ facet, value }: Facet) => {
-    setFacets((previous) => {
-      const updatedFacetGroup: FacetGroup = {
-        programmingLanguage:
-          facet === "programmingLanguage"
-            ? (value as ProgrammingLanguage)
-            : previous?.programmingLanguage,
-        category:
-          facet === "category"
-            ? (value as CodeExampleCategory)
-            : previous?.category,
-        docsSet: facet === "docsSet" ? (value as DocsSet) : previous?.docsSet,
-      };
-      return updatedFacetGroup;
-    });
-  };
+  const { results } = useAcala();
 
   return (
     <div className={styles.results_page}>
-      <header>
-        <SearchInput
-          onSubmit={(event) => {
-            handleSearch(event);
-          }}
-          aria-label="search box"
-          className={styles.search_input}
-        ></SearchInput>
-
-        <div className={styles.facet_container}>
-          <Combobox
-            label="Programming Language"
-            placeholder="Select language"
-            size="xsmall"
-            onChange={(value: string | null) => {
-              if (value) {
-                handleFacetChange({
-                  facet: "programmingLanguage",
-                  value: value,
-                });
-              }
-            }}
-          >
-            {Object.values(ProgrammingLanguageDisplayValues).map((language) => (
-              <ComboboxOption
-                key={language}
-                value={language}
-              />
-            ))}
-          </Combobox>
-
-          <Combobox
-            label="Category"
-            placeholder="Select category"
-            size="xsmall"
-            onChange={(value: string | null) => {
-              if (value) {
-                handleFacetChange({
-                  facet: "category",
-                  value: value,
-                });
-              }
-            }}
-          >
-            {Object.values(CodeExampleDisplayValues).map((category) => (
-              <ComboboxOption
-                key={category}
-                value={category}
-              />
-            ))}
-          </Combobox>
-
-          <Combobox
-            label="Documentation set"
-            placeholder="Select docs set"
-            size="xsmall"
-            onChange={(value: string | null) => {
-              if (value) {
-                handleFacetChange({
-                  facet: "docsSet",
-                  value: value,
-                });
-              }
-            }}
-          >
-            {Object.values(DocsSetDisplayValues).map((docsSet) => (
-              <ComboboxOption
-                key={docsSet}
-                value={docsSet}
-              />
-            ))}
-          </Combobox>
-        </div>
-      </header>
-
       <div className={styles.horizontal_container}>
         <div className={styles.results_container}>
           {results && (
