+++ audit/ask-cal/source/styles/main.css
@@ -61,13 +61,9 @@ body {
 }
 
 .App {
-  display: grid;
-  grid-template-areas:
-    "header      header"
-    "left-column right-column"
-    "footer      footer";
-  grid-template-rows: auto 1fr auto;
-  grid-template-columns: auto 3fr;
+  display: flex;
+  flex-direction: column;
+  align-items: center;
   height: 100vh;
 }
